+++ library/src/androidMain/kotlin/io/github/ezer_mackenzie/proximitytransfer/core/transfer/progress/CurrentTimeMillis.android.kt
@@ -1,0 +1,3 @@
+package io.github.ezer_mackenzie.proximitytransfer.core.transfer.progress
+
+internal actual fun getCurrentTimeMillis(): Long = System.currentTimeMillis()