+++ library/src/appleMain/kotlin/io/github/ezer_mackenzie/proximitytransfer/core/transfer/progress/CurrentTimeMillis.apple.kt
@@ -1,0 +1,6 @@
+package io.github.ezer_mackenzie.proximitytransfer.core.transfer.progress
+
+import platform.Foundation.NSDate
+import platform.Foundation.timeIntervalSince1970
+
+internal actual fun getCurrentTimeMillis(): Long = (NSDate().timeIntervalSince1970 * 1000).toLong()