+++ library/src/linuxX64Main/kotlin/io/github/ezer_mackenzie/proximitytransfer/core/transfer/progress/CurrentTimeMillis.linux.kt
@@ -1,0 +1,7 @@
+package io.github.ezer_mackenzie.proximitytransfer.core.transfer.progress
+
+import kotlinx.cinterop.ExperimentalForeignApi
+import platform.posix.time
+
+@OptIn(ExperimentalForeignApi::class)
+internal actual fun getCurrentTimeMillis(): Long = platform.posix.time(null) * 1000L